requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.6
executing==1.2.0
filelock==3.10.7
Flask-WTF~=1.1.1
Flask==2.3.2
Frozen-Flask==0.18
idna==3.4
iniconfig==2.0.0
ipykernel==6.23.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
openai==0.27.7
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
Pygments==2.15.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.0.2
requests==2.30.0
six==1.16.0
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
urllib3==2.0.2
virtualenv==20.21.0
wcwidth==0.2.6
Werkzeug==2.3.4
wtforms==2.3.3
